+++ projects/start-wrt/web/src/app/services/api/api.service.ts
@@ -217,6 +217,7 @@ export type SetUciRes<T extends string[]> = {
 
 export type SystemInfoRes = {
   version: string
+  gitHash: string
   language: string
   date: string
   theme: 'dark' | 'light' | 'system'